FILE: organizes internal folder into shared, transformer, and counter folders
