+++ libs/core/util-selection/src/lib/mock-data.ts
@@ -1,1 +1,1 @@
-export * from './internal/mock-data';
+export * from './shared/mock-data';
