+++ libs/core/util-selection/src/lib/shared/reduce-recursively.ts
