+++ libs/core/util-selection/src/lib/shared/types.ts
