+++ libs/core/util-selection/src/lib/transformer/ids/create-map.spec.ts
@@ -1,4 +1,4 @@
-import { AFRICA, getChildren, getId } from './mock-data';
+import { AFRICA, getChildren, getId } from '../../shared/mock-data';
 import { IdsMap, createMap } from './create-map';
 
 describe('createMap', () => {
