+++ libs/core/util-selection/src/lib/transformer/to-set.spec.ts
@@ -5,7 +5,7 @@ import {
   MOCK_IDS,
   SOME_SELECTED_SET_MODEL,
   SOME_SELECTED_STATES,
-} from './mock-data';
+} from '../shared/mock-data';
 
 describe('toSet', () => {
   it('transforms empty states', () => {
