+++ libs/core/util-selection/src/lib/transformer/to-states.ts
@@ -1,5 +1,5 @@
-import { Ids } from './ids';
-import { Model, States } from './types';
+import { Ids } from './ids/ids';
+import { Model, States } from '../shared/types';
 
 export function toStates<T>(model: Model, ids: Ids<T>): States {
   const states: States = {};
